Unset GIT_CONFIG when running commands
GIT_CONFIG is an environment variable that overrides which files `config
--get` actually fetches from. The benefit in setting this is that if you
want to add configuration, you don't have to use `--global`. I would
never recommend anyone set this, but in the case that you do, CocoaPods
can never find local specs repos by matching URLs, since getting
remote.origin.url will come from a different file, and likely always be
empty.

There's a `--local` flag that is also supposed to be the opposite of
`--global` for this, but setting this when GIT_CONFIG is set results in
an error. The other option would be to pass the file explicitly with
`--file .git/config`, but this feels more fragile.